move: extension method to another file
